move functionality to utils.py
